renamed pathch in workspace_manager_presenter.py
