+++ mslice/tests/workspace_provider_test.py
@@ -1,6 +1,5 @@
 from __future__ import (absolute_import, division, print_function)
 import numpy as np
-import mock
 from mock import patch
 import unittest
 from mslice.models.workspacemanager.workspace_algorithms import (subtract, add_workspace_runs, combine_workspace,
@@ -9,9 +8,8 @@ from mslice.models.workspacemanager.workspace_provider import (get_workspace_han
                                                                delete_workspace, rename_workspace)
 from mslice.models.workspacemanager.workspace_algorithms import processEfixed
 from mslice.util.mantid.mantid_algorithms import ConvertToMD, CloneWorkspace, CreateSimulationWorkspace
-from mslice.widgets.workspacemanager.command import Command
+
 from mantid.simpleapi import AddSampleLog
-from mslice.presenters.workspace_manager_presenter import WorkspaceManagerPresenter
 
 
 class MantidWorkspaceProviderTest(unittest.TestCase):
@@ -69,48 +67,6 @@ class MantidWorkspaceProviderTest(unittest.TestCase):
         new_ws = get_workspace_handle('newname')
         self.assertFalse(new_ws.ef_defined)
         self.assertEqual(new_ws.limits['DeltaE'], [0, 2, 1])
-
-    @patch('mslice.models.workspacemanager.workspace_provider.rename_workspace')
-    @patch('mslice.models.workspacemanager.workspace_provider.get_visible_workspace_names')
-    def test_rename_workspace(self, get_ws_names_mock, rename_ws_mock):
-        self.presenter = WorkspaceManagerPresenter(self.view)
-        # Create a view that will return a single selected workspace on call to get_workspace_selected and supply a
-        # name on call to get_workspace_new_name
-        old_workspace_name = 'file1'
-        new_workspace_name = 'new_name'
-        self.view.get_workspace_selected = mock.Mock(return_value=[old_workspace_name])
-        self.view.get_workspace_new_name = mock.Mock(return_value=new_workspace_name)
-        self.view.display_loaded_workspaces = mock.Mock()
-        get_ws_names_mock.return_value = ['file1', 'file2', 'file3']
-
-        self.presenter.notify(Command.RenameWorkspace)
-        self.view.get_workspace_selected.assert_called_once_with()
-        self.view.get_workspace_new_name.assert_called_once_with()
-        rename_ws_mock.assert_called_once_with('file1', 'new_name')
-        self.view.display_loaded_workspaces.assert_called_once()
-
-    @patch('mslice.models.workspacemanager.workspace_provider.rename_workspace')
-    def test_rename_workspace_multiple_workspace_selected_prompt_user(self, rename_ws_mock):
-        self.presenter = WorkspaceManagerPresenter(self.view)
-        # Create a view that reports multiple selected workspaces on calls to get_workspace_selected
-        selected_workspaces = ['ws1', 'ws2']
-        self.view.get_workspace_selected = mock.Mock(return_value=selected_workspaces)
-
-        self.presenter.notify(Command.RenameWorkspace)
-        self.view.get_workspace_selected.assert_called_once_with()
-        self.view.error_select_only_one_workspace.assert_called_once_with()
-        rename_ws_mock.assert_not_called()
-
-    @patch('mslice.models.workspacemanager.workspace_provider.rename_workspace')
-    def test_rename_workspace_non_selected_prompt_user(self, rename_ws_mock):
-        self.presenter = WorkspaceManagerPresenter(self.view)
-        # Create a view that reports that no workspaces are selected on calls to get_workspace_selected
-        self.view.get_workspace_selected = mock.Mock(return_value=[])
-
-        self.presenter.notify(Command.RenameWorkspace)
-        self.view.get_workspace_selected.assert_called_once_with()
-        self.view.error_select_one_workspace.assert_called_once_with()
-        rename_ws_mock.assert_not_called()
 
     def test_propagate_properties(self):
         ws_2 = CreateSimulationWorkspace(OutputWorkspace='test_ws_2', Instrument='MAR', BinParams=[-1, 1, 20],
